Optimizing for Google Page Speed (#27)
* Removed Google Fonts Links as they are not used.

Closes #21

* Added node_modules to .gitignore

* Added production mode to gulp

* Removed node_modules which should be ignored

* Now checking if wordpress is running in localhost

If so, the regualr css is loaded. Otherwise, wordpress loads the
minified version.
Also changed the gulpfile.js accordingly.

See #25

* Removed unused js scripts and minified remaining

The gulpfile now has a function to minify js. The production function is
calling 'minify-css' and 'minify-js' now instead of doing it directly.

In the footer, the minified js script is only included if worpdress is
not currently running on localhost.

close #23